touchup; split Errors into distinct file
